refactor: replace function, but trouble with the circular modules
