+++ bot.py
@@ -1,24 +1,8 @@
 import telebot
 import tokens
 import app
-import threading
-import datetime
 
 bot = telebot.TeleBot(tokens.token)
-delay = 5
-pause = bool()
-
-
-def get_prices():
-    global pause
-    if not pause:
-        threading.Timer(delay, get_prices).start()
-        now = datetime.datetime.now()
-        datum = now.strftime("%H:%M:%S")
-        print(datum)
-    else:
-        pass
-    return []
 
 
 # @handler_
@@ -29,9 +13,11 @@ def clarify_the_ticket():  # Уточнить номер билета, на сл
 
 def simulate_call():
     global pause
-    get_prices()
+    app.get_prices()
     pause = bool(input())  # False => stops algorithm
 
 
 pause = bool(input())
 simulate_call()
+
+bot.polling(none_stop=True, interval=0)
